refactor: 김치 프리미엄 리팩토링(불필요한 코드 삭제 & 에러처리 수정)
- 로직에 필요없는 파일 또는 코드 삭제
- 김치 프리미엄 에러처리
- 김치 프리미엄 반환값 OffsetResponse 사용하도록 수정

+++ src/main/java/_1danhebojo/coalarm/coalarm_service/domain/dashboard/repository/entity/KimchiPremiumEntity.java
@@ -29,7 +29,7 @@ public class KimchiPremiumEntity {
     private BigDecimal globalPrice;
 
     @Column(name = "exchange_rate", nullable = false)
-    private Integer exchangeRate;
+    private BigDecimal exchangeRate;
 
     @Column(name = "kimchi_premium", nullable = false)
     private BigDecimal kimchiPremium;
@@ -40,7 +40,7 @@ public class KimchiPremiumEntity {
     @Column(name = "reg_dt", nullable = false)
     private Instant regDt;
 
-    public KimchiPremiumEntity(CoinEntity coin, BigDecimal domesticPrice, BigDecimal globalPrice, Integer exchangeRate,
+    public KimchiPremiumEntity(CoinEntity coin, BigDecimal domesticPrice, BigDecimal globalPrice, BigDecimal exchangeRate,
                                BigDecimal kimchiPremium, BigDecimal dailyChange) {
         this.coin = coin;
         this.domesticPrice = domesticPrice;
